# docs/resultados/api_evidence/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThriveMind API",
    "description": "Ecosistema de bienestar holístico orquestado por IA",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/checkin/": {
      "post": {
        "tags": [
          "Check-in Diario"
        ],
        "summary": "Crear check-in diario",
        "operationId": "crear_checkin_api_v1_checkin__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Crear Checkin Api V1 Checkin  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/checkin/dashboard/tendencias": {
      "get": {
        "tags": [
          "Check-in Diario"
        ],
        "summary": "Obtener Tendencias Dashboard",
        "operationId": "obtener_tendencias_dashboard_api_v1_checkin_dashboard_tendencias_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 7,
              "default": 14,
              "title": "Dias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Obtener Tendencias Dashboard Api V1 Checkin Dashboard Tendencias Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkin/correlaciones": {
      "get": {
        "tags": [
          "Check-in Diario"
        ],
        "summary": "Correlaciones cross-pilar del usuario",
        "operationId": "get_correlaciones_api_v1_checkin_correlaciones_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Dias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkin/correlaciones/lag": {
      "get": {
        "tags": [
          "Check-in Diario"
        ],
        "summary": "Correlaciones con efecto retrasado",
        "operationId": "get_correlaciones_lag_api_v1_checkin_correlaciones_lag_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Dias"
            }
          },
          {
            "name": "max_lag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "Max Lag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkin/contextual": {
      "post": {
        "tags": [
          "Check-in Diario"
        ],
        "summary": "Crear Checkin Contextual",
        "operationId": "crear_checkin_contextual_api_v1_checkin_contextual_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tipo_checkin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tipo Checkin"
            }
          },
          {
            "name": "estado_emocional",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Estado Emocional"
            }
          },
          {
            "name": "energia_fisica",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Energia Fisica"
            }
          },
          {
            "name": "emocion_principal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Emocion Principal"
            }
          },
          {
            "name": "nota",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Nota"
            }
          },
          {
            "name": "hambre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Hambre"
            }
          },
          {
            "name": "saciedad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Saciedad"
            }
          },
          {
            "name": "referencia_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Referencia Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Crear Checkin Contextual Api V1 Checkin Contextual Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinCreate": {
        "properties": {
          "estado_emocional": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Estado Emocional"
          },
          "energia_fisica": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Energia Fisica"
          },
          "horas_sueno": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 24.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Horas Sueno"
          },
          "conexion_entorno": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Conexion Entorno"
          },
          "emocion_principal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emocion Principal"
          },
          "nota": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nota"
          },
          "tipo_checkin": {
            "type": "string",
            "enum": [
              "diario",
              "pre_meditacion",
              "post_meditacion",
              "pre_comida",
              "post_comida",
              "post_cosecha"
            ],
            "title": "Tipo Checkin",
            "default": "diario"
          },
          "referencia_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referencia Id"
          },
          "hambre": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hambre"
          },
          "saciedad": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Saciedad"
          },
          "hrv_estimado": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hrv Estimado"
          }
        },
        "type": "object",
        "required": [
          "estado_emocional",
          "energia_fisica",
          "conexion_entorno"
        ],
        "title": "CheckinCreate",
        "description": "Schema para crear un check-in diario."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}